refactor(#12): 단어 아이템 폴더 구조 및 파일 정리
