+++ entities/word/api/type.ts
