+++ entities/word/index.ts
@@ -1,1 +1,1 @@
-export * from './type'
+export * from './api/type'
