+++ features/word-list/index.ts
@@ -1,0 +1,2 @@
+export { default as WordItem } from './ui/WordItem'
+export { default as useWord } from './model/useWord'
