--- widgets/word-list-container/index.ts
@@ -1,2 +1,0 @@
-export { default as WordItem } from './ui/WordItem'
-export { default as WordList } from './ui/WordList'
